Stencil rejection is working, and more cleaning
