--- libraries/render-utils/src/drawOpaqueStencil.slf
@@ -1,16 +1,0 @@
-<@include gpu/Config.slh@>
-<$VERSION_HEADER$>
-//  Generated on <$_SCRIBE_DATE$>
-//
-//  drawOpaqueStencil.frag
-//  fragment shader
-//
-//  Created by Sam Gateau on 9/29/15.
-//  Copyright 2015 High Fidelity, Inc.
-//
-//  Distributed under the Apache License, Version 2.0.
-//  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
-//
-
-void main(void) {
-}
